mingle pedge_export_gv: take 'ne' as a 'size_t'
This pays down some debt from previous changes.

+++ cmd/mingle/minglemain.cpp
@@ -451,7 +451,7 @@ static int bundle(Agraph_t *g, const opts_t &opts) {
 	    	export_dot(outfile, A->m, edges, g);
 	}
 	else {
-		pedge_export_gv(outfile, (int)A->m, edges);
+		pedge_export_gv(outfile, A->m, edges);
 	}
 	return 0;
 }
